Create Arduino
In the voice recognition module, the list of four categories of garbage is pre-configured, and the serial port output value (garbage classification identifier) corresponding to each category is set. After voice recognition, the voice prompt of corresponding garbage classification is broadcast through the speaker. At the same time, the serial port of the module communicates with the serial port of the arduino uno development board to output the corresponding garbage classification identifier. The arduino uno development board is programmed to read the classification identifier input from the serial port, and then control the steering gear corresponding to the classification of garbage to rotate at a certain Angle and lift the lid. The steering gear automatically resets after 2 seconds, thus closing the lid and realizing voice garbage classification.